Span instead of raw pointers

+++ src/world/Chunk.h
@@ -19,16 +19,16 @@ inline auto Noise() noexcept -> FastNoiseLite&
 {
 	Chunk chunk;
 
-	for(uint16_t z = 0u; z < Chunk::Size; z++)
+	for(uint8_t z = 0u; z < Chunk::Size; z++)
 	{
-		for(uint16_t x = 0u; x < Chunk::Size; x++)
+		for(uint8_t x = 0u; x < Chunk::Size; x++)
 		{
 			glm::vec2 p = glm::vec2(x, z) + glm::vec2(coordinate) * static_cast<float>(Chunk::Size);
 			p /= 4.0f;
 
-			uint32_t h = ((Noise().GetNoise(p.x, p.y) + 1.0f) / 2.0f) * Chunk::Size;
+			uint8_t h = static_cast<uint8_t>(((Noise().GetNoise(p.x, p.y) + 1.0f) / 2.0f) * Chunk::Size);
 
-			for(uint16_t y = 0u; y < Chunk::Size; y++)
+			for(uint8_t y = 0u; y < Chunk::Size; y++)
 			{
 				chunk.Set(x, y, z, y <= h);
 			}
